requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
basicauth==0.4.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
bokeh==2.0.0
boto3==1.12.27
botocore==1.15.27
cchardet==2.1.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
ckanapi==4.3
click==7.1.1
colorama==0.4.3
colorlog==4.1.0
coverage==5.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
email-validator==1.0.5
et-xmlfile==1.0.1
gitdb==4.0.2
GitPython==3.1.0
hdx-python-api==4.3.5
hdx-python-country==2.4.8
hdx-python-utilities==2.2.4
html5lib==1.0.1
idna==2.9
ijson==2.6.1
importlib-metadata==1.5.0
inflect==4.1.0
injectable==2.0.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.5
jsonlines==1.2.0
jsonpath-rw==1.4.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libhxl==4.18
linear-tsv==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.2.0
ndg-httpsclient==0.5.1
num2words==0.5.10
numpy==1.18.2
openpyxl==3.0.3
packaging==20.3
pandas==1.0.3
paramiko==2.7.1
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
pprintpp==0.4.0
psycopg2-binary==2.8.4
py==1.8.1
pyaml==20.3.1
pyasn1==0.4.8
pycountry==19.8.18
pycountry-convert==0.7.2
pycparser==2.20
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
pytest-mock==2.0.0
python-chain==1.0.3
python-crontab==2.4.0
python-dateutil==2.8.1
python-io-wrapper==0.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
quantulum3==0.7.3
ratelimit==2.2.1
repoze.lru==0.7
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
smmap==3.0.1
soupsieve==2.0
SQLAlchemy==1.3.15
sshtunnel==0.1.5
tabulator==1.36.0
text-unidecode==1.3
tornado==6.0.4
tqdm==4.43.0
typing==3.7.4.1
typing-extensions==3.7.4.1
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
xlrd==1.2.0
yamlloader==0.5.5
zipp==3.1.0
